+++ mlrun/api/api/endpoints/functions.py
@@ -9,7 +9,8 @@ from sqlalchemy.orm import Session
 
 from mlrun.api.api import deps
 from mlrun.api.api.utils import log_and_raise, get_run_db_instance
-from mlrun.api.singletons import get_db, get_k8s
+from mlrun.api.utils.singletons.db import get_db
+from mlrun.api.utils.singletons.k8s import get_k8s
 from mlrun.builder import build_runtime
 from mlrun.config import config
 from mlrun.run import new_function
