+++ mlrun/api/api/endpoints/projects.py
@@ -8,7 +8,7 @@ from mlrun.api import schemas
 from mlrun.api.api import deps
 from mlrun.api.api.utils import log_and_raise
 from mlrun.api.db.sqldb.helpers import to_dict as db2dict
-from mlrun.api.singletons import get_db
+from mlrun.api.utils.singletons.db import get_db
 
 router = APIRouter()
