+++ mlrun/api/api/endpoints/tags.py
@@ -7,7 +7,7 @@ from sqlalchemy.orm import Session
 from mlrun.api.api import deps
 from mlrun.api.api.utils import log_and_raise
 from mlrun.api.db.sqldb.helpers import to_dict as db2dict, table2cls
-from mlrun.api.singletons import get_db
+from mlrun.api.utils.singletons.db import get_db
 
 router = APIRouter()
