+++ mlrun/api/api/utils.py
@@ -10,7 +10,9 @@ from sqlalchemy.orm import Session
 
 from mlrun.api import schemas
 from mlrun.api.db.sqldb.db import SQLDB
-from mlrun.api.singletons import get_db, get_logs_dir, get_scheduler
+from mlrun.api.utils.singletons.db import get_db
+from mlrun.api.utils.singletons.logs_dir import get_logs_dir
+from mlrun.api.utils.singletons.scheduler import get_scheduler
 from mlrun.config import config
 from mlrun.db.sqldb import SQLDB as SQLRunDB
 from mlrun.run import import_function, new_function
